requirements.txt
gspread==6.1.0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
